Fix: refaire une belle architecture
