+++ auth/package.json
@@ -10,9 +10,9 @@
   "description": "",
   "dependencies": {
     "crypto": "^1.0.1",
-    "dotenv": "^16.4.5",
+    "dotenv": "^16.4.6",
     "express": "^4.21.1",
-    "mongodb": "^6.10.0",
-    "mongoose": "^8.8.1"
+    "mongodb": "^6.11.0",
+    "mongoose": "^8.8.3"
   }
 }
