+++ auth/routes/authRoutes.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const { register, login } = require('../controllers/authController');
+
+const router = express.Router();
+
+router.post('/register', register);
+router.post('/login', login);
+
+module.exports = router;
