+++ magasin/package.json
@@ -10,8 +10,10 @@
   "license": "ISC",
   "description": "",
   "dependencies": {
-    "dotenv": "^16.4.5",
+    "crypto": "^1.0.1",
+    "dotenv": "^16.4.6",
     "express": "^4.21.1",
-    "mongoose": "^8.7.2"
+    "mongodb": "^6.11.0",
+    "mongoose": "^8.8.3"
   }
 }
