+++ magasin/shared/init_mongodb.js
@@ -5,7 +5,7 @@ const URI = `mongodb+srv://${process.env.MONGO_USERNAME}:${process.env.MONGO_PAS
 
 const connectDB = async () => {
     try {
-        await mongoose.connect(URI, { useNewUrlParser: true, useUnifiedTopology: true });
+        await mongoose.connect(URI);
         console.log('MongoDB connected successfully');
     } catch (error) {
         console.error('Erreur de connexion MongoDB:', error.message);
