--- shared/init_monogdb.js
@@ -1,15 +1,0 @@
-const mongoose = require('mongoose');
-require('dotenv').config();
-
-const URI = `mongodb+srv://${process.env.MONGO_USERNAME}:${process.env.MONGO_PASSWORD}@${process.env.MONGO_CLUSTER}/`;
-const connectDB = async () => {
-    try {
-        await mongoose.connect(URI);
-        console.log('MongoDB connected successfully');
-    } catch (error) {
-        console.error('Error connecting to MongoDB:', error.message);
-        process.exit(1);
-    }
-};
-
-module.exports = connectDB;
